--- app.py
@@ -1,9 +1,0 @@
-"""
-Process POST requests from clients, log data to files so it can be sent with filebeat
-@author: degenerat3
-"""
-
-from app import app
-
-if __name__ == '__main__':
-	app.run(debug=True, host='0.0.0.0')
