--- app/__init__.py
@@ -1,5 +1,0 @@
-from flask import Flask, request
-
-app = Flask(__name__)
-
-from . import views
