+++ server/app.py
@@ -1,3 +1,15 @@
+"""
+Process POST requests from clients, log data to files so it can be sent with filebeat
+@author: degenerat3
+"""
+
+import datetime
+import os
+from flask import Flask, request
+
+
+app = Flask(__name__)
+
 
 def log_fw(ip, rules):
     """
@@ -58,3 +70,29 @@ def log_arp(ip, arp):
     for entry in arp:
         print(entry)
     return
+
+
+@app.route('/campfire', methods=['POST'])
+def process_campfire():
+    content = request.json
+    ip = content['ip']
+    fwall = content['rules']
+    hosts = content['etchosts']
+    routes = content['routes']
+    arp = content['arp']
+
+    log_fw(ip, fwall)
+    #log_hosts(ip, hosts)
+    #log_routes(ip, routes)
+    #log_arp(ip, arp)
+
+    return "camped"
+
+
+@app.route('/scavpro', methods=['POST'])
+def process_scavpro():
+    content = request.json
+    return
+
+if __name__ == '__main__':
+	app.run(debug=True, host='0.0.0.0')
